Check if service exists before disabling
Certain services only exist on certain versions of windows 10. For example the `WMPNetworkSvc` does not exist in the `N` versions of Windows. This change will catch the errors generated when a service does not exist and show a warning. Services that do exist are written to the output.

BTW, thanks for putting all these scripts together. Very useful.