Added "Theme" files
Adds 2 .reg files that contain the keys responsible for changing colors in the OS. The titles explain themselves. 
The .vbs file toggles between the two .reg versions and restarts explorer.exe so that the keys stored under    HKEY_CURRENT_USER\SOFTWARE\Microsoft\Windows\CurrentVersion\ will load. 
The toggle is based on the "AppsUseLightTheme" key.

DarkMode.reg is meant as an improvement on the darktheme.reg option in the dir above as it contains a few extra tweaks and is for HKEY_CURRENT_USER only

LightMode.reg is a return to "Vanilla" Windows color settings.